backend/requirements.txt
Flask==3.0.0
ldap3==2.9.1
better-sqlite3==9.4.3
python-dotenv==1.0.0
requests==2.31.0
